Added test and more logic.

+++ Loodsen.SalaryCalculator/Models/Payment.cs
@@ -2,5 +2,5 @@
 
 public readonly record struct Payment(
     decimal Value,
-    float LossesPercent,
+    decimal LossesPercent,
     DateOnly Date);
